+++ src/main/java/com/banquito/core/general/modelo/Sucursal.java
@@ -47,8 +47,8 @@ public class Sucursal {
     @Column(name = "estado", nullable = false, length = 15)
     private String estado;
 
-    @Column(name = "version", nullable = false, precision = 9)
-    private BigDecimal version;
+    @Column(name = "version", nullable = false)
+    private Long version;
 
     // Getters y Setters
     public String getCodigo() {
@@ -147,11 +147,11 @@ public class Sucursal {
         this.estado = estado;
     }
 
-    public BigDecimal getVersion() {
+    public Long getVersion() {
         return version;
     }
 
-    public void setVersion(BigDecimal version) {
+    public void setVersion(Long version) {
         this.version = version;
     }
